util-lib: move inotify-related definitions to fs-util.[ch]
